+++ packages/ebsr/.npmignore
